user-behavior-scenarios/registrator/requirements.txt
Appium-Python-Client==2.2.0
asgiref==3.5.0
async-generator==1.10
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
cryptography==36.0.2
frida==15.1.17
frida-tools==10.6.1
h11==0.13.0
idna==3.3
outcome==1.1.0
prompt-toolkit==3.0.29
pycparser==2.21
Pygments==2.12.0
pyOpenSSL==22.0.0
PySocks==1.7.1
pytz==2022.1
requests==2.27.1
selenium==4.1.3
sentry-sdk==1.9.8
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
trio==0.20.0
trio-websocket==0.9.2
tzdata==2022.1
urllib3==1.26.12
wcwidth==0.2.5
wsproto==1.1.0
